thx for your work

+++ src/server/game/Handlers/SpellHandler.cpp
@@ -337,7 +337,7 @@ void WorldSession::HandleCancelAuraOpcode(WorldPackets::Spells::CancelAura& canc
     if (spellInfo->IsChanneled())
     {
         if (Spell* curSpell = _player->GetCurrentSpell(CURRENT_CHANNELED_SPELL))
-            if (curSpell->GetSpellInfo()->Id == cancelAura.SpellID)
+            if (curSpell->GetSpellInfo()->Id == uint32(cancelAura.SpellID))
                 _player->InterruptSpell(CURRENT_CHANNELED_SPELL);
         return;
     }
